20240814-제자리 회전 완료, 일자 이동 작업중

+++ Source/WarHorizon_AllBigGun/Aircraft/WHAircraftsBase.cpp
@@ -78,7 +78,7 @@ void AWHAircraftsBase::MoveFront()
 	AircraftMovementComp->AddInputVector(GetActorForwardVector());
 }
 
-void AWHAircraftsBase::Turn(float Angle)
+void AWHAircraftsBase::Turn(int Angle)
 {
 	AircraftMovementComp->Turn(Angle);
 }
