+++ Source/WarHorizon_AllBigGun/Interface/AircraftsInterface.h
@@ -28,7 +28,7 @@ public:
 	virtual void DestroyAircraft(int Index) = 0;
 	virtual void IncreaseHeight(float MaxZ) = 0;
 	virtual void DecreaseHeight(float MinZ) = 0;
-	virtual void Turn(float Angle) = 0;
+	virtual void Turn(int Angle) = 0;
 	virtual void SetIsSpawnEnd() = 0;
 
 	virtual void Fire() = 0;
